Add static library support
Also fixes building libSvtAv1Dec.so.  The generated so file was
essentially a stub with none of the support libs bundled.

There were also dependencies on EbSvtAv1Enc.h in SvtAv1Dec and
missing variable definitions that had to be fixed before libSvtAv1Dec.so
and SvtAv1Dec would link.